SmoothedParticleHydrodynamics

Documentation for SmoothedParticleHydrodynamics.

The fluid is represented by a discrete particles centred at the location $\mathbf r_i$:

\[\frac{d \mathbf r_i}{dt} = \mathbf v_i\]

where $\mathbf v_i$ is the velocity of the i-th particle.

SmoothedParticleHydrodynamics.KernelSpiky

SmoothedParticleHydrodynamics.Location

SmoothedParticleHydrodynamics.KernelSpiky — Method

k = KernelSpiky(n,h)

Kernel in n dimension with characteristic radius h given by the following expression:

\[W(r,h) = c \; (h - r)^3\]

for $r \le h$ where $r$ is the radial distance and where the normalization coefficient $c$ is given by:

\[\frac{1}{S_n \, c} = \int_0^h\left(h-r\right)^3 r^{n-1}dr =\frac{h^{n+3}}{n}-\frac{3h^{n+3}}{n+1}+\frac{3h^{n+3}}{n+2}-\frac{h^{n+3}}{n+3}\]

where $S_n$ is the surface of the unit n-sphere.

SmoothedParticleHydrodynamics.Location — Type

loc = Location(p)

Adapter type for SpatialHashing such that loc[i] returns p.x[i] where p is a list of particles.